requirements.txt
uvicorn
fastapi
mysql-connector-python
